+++ apps/api/src/schema/projects.ts
@@ -1,0 +1,22 @@
+import { type Static, Type } from "@sinclair/typebox";
+import { nullableStringSchema } from "./shared";
+
+// Projects
+const readProjectParamsSchema = Type.Object({
+  slug: Type.String(),
+});
+
+export const readProjectReplySchema = Type.Object({
+  slug: Type.String(),
+  name: Type.String(),
+  backendWallets: Type.Array(Type.String()),
+  callTargets: Type.Array(Type.String()),
+  redirectUris: Type.Array(Type.String()),
+  customAuth: Type.Boolean(),
+  icon: nullableStringSchema,
+  cover: nullableStringSchema,
+  color: nullableStringSchema,
+});
+
+export type ReadProjectParams = Static<typeof readProjectParamsSchema>;
+export type ReadProjectReply = Static<typeof readProjectReplySchema>;
